requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
python-socketio>=5.10.0
python-dotenv>=1.0.0
anthropic>=0.34.0
gevent>=23.9.1
gevent-websocket>=0.10.1
requests>=2.31.0
